Add: Tag & Like type to types/General dir, and refactoring files that contain those types
