+++ pages/search/_id.vue
@@ -165,15 +165,7 @@ import { db } from '@/plugins/firebase'
 import { Kuchikomi, User, Class } from '@/types/State'
 import { setAvgRating } from '@/helpers/setAvgRating'
 import { KUCHIKOMI_TAGS } from '@/data/TAGS'
-
-interface Like {
-  docId: string
-  classId: string
-  kuchikomiId: string
-  likedBy: string
-  likeTo: string
-  createdAt: string
-}
+import { Like } from '@/types/General'
 
 export default defineComponent({
   name: 'SearchId',
